+++ web_google_maps/static/src/js/widgets/gplaces_autocomplete.js
@@ -5,8 +5,9 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
     const core = require('web.core');
     const Utils = require('web_google_maps.Utils');
     const _t = core._t;
+    const GoogleMapLoaderUtil = require('web_google_maps.GoogleMapLoaderUtil');
 
-    const GplaceAutocomplete = BasicFields.InputField.extend({
+    const GplaceAutocomplete = BasicFields.InputField.extend(GoogleMapLoaderUtil, {
         tagName: 'span',
         supportedFieldTypes: ['char'],
         /**
@@ -72,11 +73,7 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                 // update 'fillfields', 'component_form', 'delimiter' if exists
                 if (this.attrs.options) {
                     if (this.attrs.options.hasOwnProperty('component_form')) {
-                        this.component_form = _.defaults(
-                            {},
-                            this.attrs.options.component_form,
-                            this.component_form
-                        );
+                        this.component_form = _.defaults({}, this.attrs.options.component_form, this.component_form);
                     }
                     if (this.attrs.options.hasOwnProperty('delimiter')) {
                         this.fillfields_delimiter = _.defaults(
@@ -95,11 +92,7 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                         if (this.force_override) {
                             this.address_form = this.attrs.options.address_form;
                         } else {
-                            this.address_form = _.defaults(
-                                {},
-                                this.attrs.options.address_form,
-                                this.address_form
-                            );
+                            this.address_form = _.defaults({}, this.attrs.options.address_form, this.address_form);
                         }
                     }
                     if (Object.prototype.hasOwnProperty.call(this.attrs.options, 'display_name')) {
@@ -108,9 +101,12 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                 }
 
                 this.target_fields = this.getFillFieldsType();
-                const self = await this.initGplacesAutocomplete();
-                self._geolocate();
+                // async function to load Google script
+                this.handleGoogleMapLoader();
             }
+        },
+        initializeGoogle: function () {
+            this.initGplacesAutocomplete().then(() => this._geolocate.bind(this));
         },
         /**
          * Geolocate
@@ -234,17 +230,17 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
          * https://developers.google.com/maps/documentation/javascript/places-autocomplete#specify-data-fields
          * @returns Array
          */
-         get_google_fields_restriction:  function() {
-            return []
-        }
+        get_google_fields_restriction: function () {
+            return [];
+        },
     });
 
     const GplaceAddressAutocompleteField = GplaceAutocomplete.extend({
         className: 'o_field_char o_field_google_address_autocomplete',
         /**
-        * @param {*} values
-        */
-       _onUpdateWidgetFields: function (values) {
+         * @param {*} values
+         */
+        _onUpdateWidgetFields: function (values) {
             values = typeof values !== 'undefined' ? values : {};
             if (values.hasOwnProperty(this.lat) || values.hasOwnProperty(this.lng)) {
                 const geometry = {
@@ -287,7 +283,7 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                 [this.address_form.country_id]: 'country',
             };
             // possible value: `address_format` or `no_address_format`
-            // address_format: widget will populate address returned by Google to Odoo address fields 
+            // address_format: widget will populate address returned by Google to Odoo address fields
             // no_address_format: no populate address, will take address and the geolocation data.
             this.address_mode = 'address_format';
             // Autocomplete request types
@@ -311,11 +307,7 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                     if (this.force_override) {
                         this.fillfields = this.attrs.options.fillfields;
                     } else {
-                        this.fillfields = _.defaults(
-                            {},
-                            this.attrs.options.fillfields,
-                            this.fillfields
-                        );
+                        this.fillfields = _.defaults({}, this.attrs.options.fillfields, this.fillfields);
                     }
                 }
                 if (Object.prototype.hasOwnProperty.call(this.attrs.options, 'types')) {
@@ -367,24 +359,15 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
          */
         populateAddress: async function (place, parse_address) {
             const requests = [];
-            let index_of_state = _.findIndex(
-                this.target_fields,
-                (f) => f.name === this.address_form.state_id
-            );
+            let index_of_state = _.findIndex(this.target_fields, (f) => f.name === this.address_form.state_id);
             const target_fields = this.target_fields.slice();
-            const field_state =
-                index_of_state > -1 ? target_fields.splice(index_of_state, 1)[0] : false;
+            const field_state = index_of_state > -1 ? target_fields.splice(index_of_state, 1)[0] : false;
 
             _.each(target_fields, (field) => {
-                requests.push(
-                    this._prepareValue(field.relation, field.name, parse_address[field.name])
-                );
+                requests.push(this._prepareValue(field.relation, field.name, parse_address[field.name]));
             });
             // Set geolocation
-            const partner_geometry = this._setGeolocation(
-                place.geometry.location.lat(),
-                place.geometry.location.lng()
-            );
+            const partner_geometry = this._setGeolocation(place.geometry.location.lat(), place.geometry.location.lng());
             _.each(partner_geometry, (val, field) => {
                 requests.push(this._prepareValue(false, field, val));
             });
@@ -415,12 +398,7 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
          * @return {Object}
          */
         get_google_fields_restriction: function () {
-            return [
-                'address_components',
-                'name',
-                'geometry',
-                'formatted_address',
-            ];
+            return ['address_components', 'name', 'geometry', 'formatted_address'];
         },
         /**
          * Override
@@ -431,22 +409,16 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                 setTimeout(() => {
                     if (!this.places_autocomplete) {
                         const google_fields = this.get_google_fields_restriction();
-                        this.places_autocomplete = new google.maps.places.Autocomplete(
-                            this.$input.get(0),
-                            {
-                                types: this.autocomplete_types,
-                                fields: google_fields,
-                            }
-                        );
+                        this.places_autocomplete = new google.maps.places.Autocomplete(this.$input.get(0), {
+                            types: this.autocomplete_types,
+                            fields: google_fields,
+                        });
                         if (this.autocomplete_settings) {
                             this.places_autocomplete.setOptions(this.autocomplete_settings);
                         }
                     }
                     // When the user selects an address from the dropdown, populate the address fields in the form.
-                    this.places_autocomplete.addListener(
-                        'place_changed',
-                        this.handlePopulateAddress.bind(this)
-                    );
+                    this.places_autocomplete.addListener('place_changed', this.handlePopulateAddress.bind(this));
                     resolve(this);
                 }, 300);
             });
@@ -463,15 +435,9 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                 if (this.lat && this.lng) {
                     fields = fields.concat([this.lat, this.lng]);
                 }
-                unknown_fields = _.filter(
-                    fields,
-                    (field) => !this.record.fields.hasOwnProperty(field)
-                );
+                unknown_fields = _.filter(fields, (field) => !this.record.fields.hasOwnProperty(field));
             } else if (this.address_mode === 'no_address_format' && this.lat && this.lng) {
-                unknown_fields = _.filter(
-                    [this.lat, this.lng],
-                    (field) => !this.record.fields.hasOwnProperty(field)
-                );
+                unknown_fields = _.filter([this.lat, this.lng], (field) => !this.record.fields.hasOwnProperty(field));
             }
             if (unknown_fields && unknown_fields.length > 0) {
                 this.do_warn(
@@ -631,19 +597,11 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
             if (place.hasOwnProperty('address_components')) {
                 const values = {};
                 const requests = [];
-                const index_of_state = _.findIndex(
-                    this.target_fields,
-                    (f) => f.name === this.address_form.state_id
-                );
+                const index_of_state = _.findIndex(this.target_fields, (f) => f.name === this.address_form.state_id);
                 const target_fields = this.target_fields.slice();
-                const field_state =
-                    index_of_state > -1 ? target_fields.splice(index_of_state, 1)[0] : false;
+                const field_state = index_of_state > -1 ? target_fields.splice(index_of_state, 1)[0] : false;
                 // Get address
-                const google_address = this._populateAddress(
-                    place,
-                    this.fillfields.address,
-                    this.fillfields_delimiter
-                );
+                const google_address = this._populateAddress(place, this.fillfields.address, this.fillfields_delimiter);
                 _.extend(values, google_address);
                 // Get place (name, website, phone)
                 const google_place = this._populatePlaces(place, this.fillfields.general);
@@ -656,9 +614,7 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                 _.extend(values, google_geolocation);
 
                 _.each(target_fields, (field) => {
-                    requests.push(
-                        this._prepareValue(field.relation, field.name, values[field.name])
-                    );
+                    requests.push(this._prepareValue(field.relation, field.name, values[field.name]));
                 });
 
                 const result = await Promise.all(requests);
@@ -701,22 +657,16 @@ odoo.define('web_google_maps.GplaceAutocompleteFields', function (require) {
                 setTimeout(() => {
                     if (!this.places_autocomplete) {
                         const google_fields = this.get_google_fields_restriction();
-                        this.places_autocomplete = new google.maps.places.Autocomplete(
-                            this.$input.get(0),
-                            {
-                                types: this.autocomplete_types,
-                                fields: google_fields,
-                            }
-                        );
+                        this.places_autocomplete = new google.maps.places.Autocomplete(this.$input.get(0), {
+                            types: this.autocomplete_types,
+                            fields: google_fields,
+                        });
                         if (this.autocomplete_settings) {
                             this.places_autocomplete.setOptions(this.autocomplete_settings);
                         }
                     }
                     // When the user selects an address from the dropdown, populate the address fields in the form.
-                    this.places_autocomplete.addListener(
-                        'place_changed',
-                        this.handlePopulateAddress.bind(this)
-                    );
+                    this.places_autocomplete.addListener('place_changed', this.handlePopulateAddress.bind(this));
                     resolve(this);
                 }, 300);
             });